requirements.txt
selenium==4.15.2
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
